Fix sign conversion warnings in SQL connection and statement classes
Add explicit casts for SQLLEN/SQLSMALLINT to size_t conversions:
- SqlConnection.cpp: Cast nameLen/textLen in database info retrieval
- SqlConnectInfo.cpp: Cast ranges::distance result to size_t
- SqlStatement.cpp: Cast parameter count and row/column counts
- SqlVariant.cpp: Add NOLINT for cognitive complexity in GetColumn
- UnicodeConverter.cpp: Cast bitwise operation result to char32_t

+++ src/Lightweight/SqlConnectInfo.cpp
@@ -61,7 +61,7 @@ std::string SqlConnectionString::SanitizePwd(std::string_view input)
 SqlConnectionStringMap ParseConnectionString(SqlConnectionString const& connectionString)
 {
     auto pairs = connectionString.value | std::views::split(';') | std::views::transform([](auto pair_view) {
-                     return std::string_view(&*pair_view.begin(), std::ranges::distance(pair_view));
+                     return std::string_view(&*pair_view.begin(), static_cast<size_t>(std::ranges::distance(pair_view)));
                  });
 
     SqlConnectionStringMap result;
